Bug fix; Add support for TDK kHz range
The gyrometer axes should be disabled for this project, this commit fixes
that. (No actual functional change to firmware).

This allows for setting the sample rate to 1KHz and greater for the TDK
sensor. This introduces a new config variable SNSR_SAMPLE_RATE_UNIT which
must be set to SNSR_SAMPLE_RATE_UNIT_HZ or SNSR_SAMPLE_RATE_UNIT_KHZ.
This was the only way to do things consistently across both drivers
using preprocessor macros.